requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.4.2
backpack==0.1
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
cleo==0.6.8
cryptography==2.3.1
Django==2.1.2
django-cors-headers==2.4.0
django-extensions==2.1.4
djangorestframework==3.9.0
Faker==0.8.18
greenlet==0.4.13
idna==2.7
inflection==0.3.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
orator==0.9.8
pastel==0.1.0
pendulum==1.5.1
pipenv==2018.10.13
pyaml==16.12.2
pycodestyle==2.4.0
pycparser==2.19
Pygments==2.2.0
pylev==1.3.0
pylint==2.1.1
PyMySQL==0.9.2
python-dateutil==2.7.5
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
readline==6.2.4.1
simplejson==3.16.0
six==1.11.0
text-unidecode==1.2
tzlocal==1.5.1
virtualenv==16.0.0
virtualenv-clone==0.4.0
wrapt==1.10.11
